+++ HISTORY.rst
@@ -2,7 +2,15 @@
 History
 =======
 
+1.1.0 (2020-05-12)
+------------------
+
+* Change brightness from int (0 - 255) to float (0.0 - 1.0)
+* Added logic to fetch device status on startup
+* Bugfixes
+
+
 1.0.0 (2020-05-08)
 ------------------
 
-* Initial version.
+* Initial version